refactor: Move CodeXPDialog to views package
